requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coverage==4.5.2
cryptography==2.4.2
idna==2.8
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
numpy==1.15.4
oauthlib==2.1.0
paramiko==2.4.2
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
pylint==2.2.2
pymongo==3.7.2
PyNaCl==1.3.0
PySocks==1.6.8
pytest==4.0.2
pytest-cov==2.6.1
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
SQLAlchemy==1.2.15
tinydb==3.12.2
tweepy==3.7.0
urllib3==1.24.1
wrapt==1.10.11
